chore: move models outside the command_line directory
