+++ shamba/model/crop_model.py
@@ -6,7 +6,7 @@ from typing import Tuple, List
 import numpy as np
 from marshmallow import Schema, fields, post_load
 
-from ..common import csv_handler
+from .common import csv_handler
 from .common_schema import OutputSchema as ClimateDataOutputSchema
 
 # from .crop_params import CropParams
