+++ shamba/model/emit.py
