+++ shamba/model/litter.py
@@ -8,8 +8,7 @@ import logging as log
 import numpy as np
 from marshmallow import Schema, fields, post_load
 
-from .. import configuration
-from ..common import csv_handler
+from .common import csv_handler
 from .common_schema import OutputSchema as LitterDataOutputSchema
 
 """
