+++ shamba/model/main.py
@@ -1,3 +1,4 @@
+from typing import List, Dict, Any, Optional
 from toolz import get
 
 import model.common.constants as CONSTANTS
@@ -5,7 +6,7 @@ from model.common.calculate_emissions import handle_intervention
 
 
 def run(project_name, data):
-    inputs = get("inputs", data) or None
+    inputs: Optional[List[Dict[str, Any]]] = get("inputs", data) or None # type: ignore
     allometry = str(get(CONSTANTS.ALLOMETRY_KEY, data, "chave dry"))
 
     if inputs is None:
