+++ shamba/model/soil_models/roth_c/inverse_roth_c.py
@@ -4,7 +4,7 @@ import numpy as np
 from marshmallow import fields, post_load
 from scipy import optimize
 
-from ....common import csv_handler
+from ...common import csv_handler
 from .roth_c import RothCData, RothCSchema
 from .roth_c import create as create_roth_c
 from .roth_c import dC_dt
