+++ shamba/model/tree_model.py
@@ -9,8 +9,8 @@ import numpy as np
 from tabulate import tabulate
 from marshmallow import Schema, fields, post_load
 
-from .. import configuration
-from ..common import csv_handler
+from . import configuration
+from .common import csv_handler
 from .common_schema import OutputSchema as ClimateDataOutputSchema
 from .tree_growth import TreeGrowthSchema, derivative_functions
 from .tree_params import ROOT_IN_TOP_30, TreeParamsSchema
