+++ shamba/tests/test_fertiliser_emissions.py
@@ -1,10 +1,10 @@
 import os  # Add the parent directory to the Python path
-import model.command_line.emit as Emit
+import model.emit as Emit
 import numpy as np
 import pytest
 from model.common import csv_handler
 from model import configuration
-import model.command_line.litter as LitterModel
+import model.litter as LitterModel
 
 
 def test_fertiliser_model():
