+++ shamba/tests/test_fire_emissions.py
@@ -1,5 +1,5 @@
 import os  # Add the parent directory to the Python path
-import model.command_line.emit as Emit
+import model.emit as Emit
 import numpy as np
 import pytest
 from model.common import csv_handler
